Move swimlane validator methods
